PluginManager: Move plugin manager into its own project
The plugin manager now is its own subproject and not part
of the core anymore and will be the base for further refactoring.

Requires some changes in the packaging to include the jar as
well as some minor test modifications.
