+++ core/src/test/java/org/elasticsearch/test/hamcrest/ElasticsearchFileAssertions.java
@@ -1,0 +1,53 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.test.hamcrest;
+
+import org.junit.Assert;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertFileExists;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+
+public class ElasticsearchFileAssertions {
+
+    /**
+     * Check that a file contains a given String
+     * @param dir root dir for file
+     * @param filename relative path from root dir to file
+     * @param expected expected content (if null, we don't expect any file)
+     */
+    public static void assertFileContent(Path dir, String filename, String expected) throws IOException {
+        Assert.assertThat(Files.exists(dir), is(true));
+        Path file = dir.resolve(filename);
+        if (expected == null) {
+            Assert.assertThat("file [" + file + "] should not exist.", Files.exists(file), is(false));
+        } else {
+            assertFileExists(file);
+            String fileContent = new String(Files.readAllBytes(file), StandardCharsets.UTF_8);
+            // trim the string content to prevent different handling on windows vs. unix and CR chars...
+            Assert.assertThat(fileContent.trim(), equalTo(expected.trim()));
+        }
+    }
+}
