+++ distribution/src/main/assemblies/common-bin.xml
@@ -34,6 +34,7 @@
             <useTransitiveDependencies>false</useTransitiveDependencies>
             <includes>
                 <include>org.elasticsearch:elasticsearch</include>
+                <include>org.elasticsearch:elasticsearch-plugin-manager</include>
             </includes>
         </dependencySet>
     </dependencySets>
