+++ plugin-manager/src/main/java/org/elasticsearch/plugins/PluginManager.java
