+++ plugin-manager/src/test/java/org/elasticsearch/plugins/PluginManagerCliTests.java
