+++ plugin-manager/src/test/java/org/elasticsearch/plugins/PluginManagerUnitTests.java
